Update EnvironmentVariableCreate for multiple environments, add tests
