+++ packages/eas-cli/src/graphql/types/EnvironmentVariable.ts
@@ -5,8 +5,9 @@ export const EnvironmentVariableFragmentNode = gql`
     id
     name
     value
-    environment
+    environments
     createdAt
+    updatedAt
     scope
     visibility
   }
