--- packages/eas-cli/src/utils/formatVariable.ts
@@ -1,14 +1,0 @@
-import dateFormat from 'dateformat';
-
-import formatFields from './formatFields';
-import { EnvironmentVariableFragment } from '../graphql/generated';
-
-export function formatVariable(variable: EnvironmentVariableFragment): string {
-  return formatFields([
-    { label: 'ID', value: variable.id },
-    { label: 'Name', value: variable.name },
-    { label: 'Value', value: variable.value ?? '(secret)' },
-    { label: 'Scope', value: variable.scope },
-    { label: 'Created at', value: dateFormat(variable.createdAt, 'mmm dd HH:MM:ss') },
-  ]);
-}
